+++ app/src/main/java/com/sather/todo/ui/navigation/ToDoNavHost.kt
@@ -81,7 +81,7 @@ fun ToDoNavHost(
                 route = RoutineEntryDestination.routeWithArgs,
                 arguments = RoutineEntryDestination.arguments
             ) {
-                val backlogId=it.arguments?.getInt("backlogId")?:0
+                val backlogId=it.arguments?.getLong("backlogId")?:0
                 RoutineEntryScreen(
                     backlogId =backlogId,
                     navigateBack= { navController.popBackStack() }
